Make execution_span a composite module and add docs
